Remove last few v2 playwright pages
