+++ e2e_playwright/pages/manage/addBedspacePage.ts
@@ -1,6 +1,6 @@
 import { Page, expect } from '@playwright/test'
 import { Bedspace } from '@temporary-accommodation-ui/e2e'
-import { BasePage } from '../../basePage'
+import { BasePage } from '../basePage'
 
 export class AddBedspacePage extends BasePage {
   static async initialise(page: Page) {
