+++ e2e_playwright/steps/manage.ts
@@ -1,22 +1,22 @@
 import { Page, expect } from '@playwright/test'
 import { Bedspace, Property } from '@temporary-accommodation-ui/e2e'
 import { DateFormats } from '../../server/utils/dateUtils'
-import { ListPropertiesPage } from '../pages/manage/v2/listPropertiesPage'
-import { AddPropertyPage } from '../pages/manage/v2/addPropertyPage'
-import { ViewPropertyPage } from '../pages/manage/v2/viewPropertyPage'
-import { EditPropertyPage } from '../pages/manage/v2/editPropertyPage'
-import { ViewBedspacePage } from '../pages/manage/v2/viewBedspacePage'
-import { AddBedspacePage } from '../pages/manage/v2/addBedspacePage'
-import { ViewBedspacesPage } from '../pages/manage/v2/viewBedspacesPage'
-import { EditBedspacePage } from '../pages/manage/v2/editBedspacePage'
-import { ArchivePropertyPage } from '../pages/manage/v2/archivePropertyPage'
-import { UnarchivePropertyPage } from '../pages/manage/v2/unarchivePropertyPage'
-import { CancelArchivePropertyPage } from '../pages/manage/v2/cancelArchivePropertyPage'
-import { ArchiveBedspacePage } from '../pages/manage/v2/archiveBedspacePage'
-import { UnarchiveBedspacePage } from '../pages/manage/v2/unarchiveBedspacePage'
-import { CancelArchiveBedspacePage } from '../pages/manage/v2/cancelArchiveBedspacePage'
-import { CancelUnarchiveBedspacePage } from '../pages/manage/v2/cancelUnarchiveBedspacePage'
-import { CancelUnarchivePropertyPage } from '../pages/manage/v2/cancelUnarchivePropertyPage'
+import { ListPropertiesPage } from '../pages/manage/listPropertiesPage'
+import { AddPropertyPage } from '../pages/manage/addPropertyPage'
+import { ViewPropertyPage } from '../pages/manage/viewPropertyPage'
+import { EditPropertyPage } from '../pages/manage/editPropertyPage'
+import { ViewBedspacePage } from '../pages/manage/viewBedspacePage'
+import { AddBedspacePage } from '../pages/manage/addBedspacePage'
+import { ViewBedspacesPage } from '../pages/manage/viewBedspacesPage'
+import { EditBedspacePage } from '../pages/manage/editBedspacePage'
+import { ArchivePropertyPage } from '../pages/manage/archivePropertyPage'
+import { UnarchivePropertyPage } from '../pages/manage/unarchivePropertyPage'
+import { CancelArchivePropertyPage } from '../pages/manage/cancelArchivePropertyPage'
+import { ArchiveBedspacePage } from '../pages/manage/archiveBedspacePage'
+import { UnarchiveBedspacePage } from '../pages/manage/unarchiveBedspacePage'
+import { CancelArchiveBedspacePage } from '../pages/manage/cancelArchiveBedspacePage'
+import { CancelUnarchiveBedspacePage } from '../pages/manage/cancelUnarchiveBedspacePage'
+import { CancelUnarchivePropertyPage } from '../pages/manage/cancelUnarchivePropertyPage'
 
 export const visitListPropertiesPage = async (page: Page) => {
   // TODO: navigate to the list properties page from the dashboard once the v2 pages are live in prod
